move websocket to pkg, no app specific logic exists
